moved DBConn class to its own folder
